+++ broker-auth-multiuser-otp/src/main/java/org/aktin/broker/auth/otp/controller/AuthEndpoint.java
@@ -1,4 +1,4 @@
-package org.aktin.broker.auth.otp.http;
+package org.aktin.broker.auth.otp.controller;
 
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
@@ -13,15 +13,15 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
-import org.aktin.broker.auth.otp.auth.Token;
-import org.aktin.broker.auth.otp.auth.TokenManager;
+import org.aktin.broker.auth.otp.token.Token;
+import org.aktin.broker.auth.otp.token.TokenManager;
 import org.aktin.broker.auth.otp.service.UserAuthService;
 import org.aktin.broker.rest.Authenticated;
 import org.aktin.broker.rest.RequireAdmin;
 import org.aktin.broker.server.auth.HttpBearerAuthentication;
 
 // TODO add proper logging of user actions
-@Path("auth/v2/")
+@Path("auth")
 public class AuthEndpoint {
 
   private static final Logger log = Logger.getLogger(AuthEndpoint.class.getName());
